+++ plugins/pi4j-plugin-linuxfs/src/main/java/com/pi4j/plugin/linuxfs/internal/LinuxPwm.java
@@ -1,4 +1,4 @@
-package com.pi4j.plugin.linuxfs.provider.pwm;
+package com.pi4j.plugin.linuxfs.internal;
 
 /*-
  * #%L
