+++ plugins/pi4j-plugin-linuxfs/src/main/java/module-info.java
@@ -37,6 +37,7 @@ module com.pi4j.plugin.linuxfs {
 
     exports com.pi4j.plugin.linuxfs;
     exports com.pi4j.plugin.linuxfs.provider.gpio.digital;
+    exports com.pi4j.plugin.linuxfs.provider.pwm;
     exports com.pi4j.plugin.linuxfs.provider.i2c;
 
     provides com.pi4j.extension.Plugin
